src/main/java/racingcar/model/Car.java
package racingcar.model;

public class Car {
    private final String name;
    private int position;

    public Car(String name) {
        this.name = name;
    }

    public String getName() {
        return name;
    }

    public int getPosition() {
        return position;
    }

    public void moveFoward() {
        position++;
    }

    public String getStatus() {
        return name + " : " + "-".repeat(position);
    }

    public void validateNotExceedMaxLength() {
        if (name.length() > 5) {
            throw new IllegalArgumentException("[ERROR] 자동차 이름은 5자 이하여야 합니다.");
        }
    }
}
